+++ taiga/importers/exceptions.py
@@ -1,0 +1,8 @@
+class InvalidRequest(Exception):
+    pass
+
+class InvalidAuthResult(Exception):
+    pass
+
+class FailedRequest(Exception):
+    pass